refactor(anvil): 重构配方缓存系统
- 新增抽象缓存元素类、缓存操作记录类等基础结构
- 优化 HasCauldron 预测逻辑，支持空炼药锅和流体转换
- 修改 BulgingCategory 渲染逻辑，适应新配方结构
- 重构 HasItemBase 和 HasItemIngredient，使用新的缓存接口
- 新增缓存输入输出实现类，支持复杂的物品操作
