+++ src/main/java/dev/dubhe/anvilcraft/recipe/anvil/cache/item/operation/CacheOperation.java
@@ -1,0 +1,7 @@
+package dev.dubhe.anvilcraft.recipe.anvil.cache.item.operation;
+
+/**
+ * 缓存操作记录类
+ */
+public record CacheOperation(int amount) {
+}
