+++ src/main/java/dev/dubhe/anvilcraft/recipe/anvil/cache/package-info.java
@@ -1,0 +1,7 @@
+@MethodsReturnNonnullByDefault
+@ParametersAreNonnullByDefault
+package dev.dubhe.anvilcraft.recipe.anvil.cache;
+
+import net.minecraft.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;
