+++ src/main/java/dev/dubhe/anvilcraft/recipe/anvil/predicate/item/HasItemBase.java
@@ -7,6 +7,7 @@ import dev.dubhe.anvilcraft.recipe.anvil.IRecipePredicate;
 import dev.dubhe.anvilcraft.recipe.anvil.InWorldRecipeContext;
 import dev.dubhe.anvilcraft.recipe.anvil.InWorldRecipeData;
 import dev.dubhe.anvilcraft.recipe.anvil.cache.ItemCache;
+import dev.dubhe.anvilcraft.recipe.anvil.cache.item.ICacheInput;
 import dev.dubhe.anvilcraft.recipe.anvil.util.IItemStackPredicate;
 import dev.dubhe.anvilcraft.util.RecipeUtil;
 import lombok.Getter;
@@ -65,9 +66,9 @@ public abstract class HasItemBase<T extends HasItemBase<T, P>, P extends IItemSt
      * @param context 配方上下文
      * @return 物品缓存输入
      */
-    public ItemCache.ICacheInput getItem(@NotNull InWorldRecipeContext context) {
+    public ICacheInput getItem(@NotNull InWorldRecipeContext context) {
         context.computeIfAbsent(ItemCache.ITEM_CACHE);
-        final InWorldRecipeData<ItemCache.ICacheInput> cacheInput = InWorldRecipeData.of(
+        final InWorldRecipeData<ICacheInput> cacheInput = InWorldRecipeData.of(
             AnvilCraft.of("item_cache_input/%s".formatted(this.hashCode())),
             (ctx, key) -> {
                 ItemCache itemCache = ctx.get(ItemCache.ITEM_CACHE);
