+++ src/main/java/dev/dubhe/anvilcraft/recipe/anvil/predicate/item/HasItemIngredient.java
@@ -5,6 +5,7 @@ import dev.dubhe.anvilcraft.init.ModRecipePredicateTypes;
 import dev.dubhe.anvilcraft.recipe.anvil.IRecipePredicate;
 import dev.dubhe.anvilcraft.recipe.anvil.InWorldRecipeContext;
 import dev.dubhe.anvilcraft.recipe.anvil.cache.ItemCache;
+import dev.dubhe.anvilcraft.recipe.anvil.cache.item.ICacheInput;
 import dev.dubhe.anvilcraft.recipe.anvil.util.ItemIngredientPredicate;
 import lombok.Getter;
 import net.minecraft.advancements.critereon.ItemSubPredicate;
@@ -42,13 +43,13 @@ public class HasItemIngredient extends HasItemBase<HasItemIngredient, ItemIngred
 
     @Override
     public void snapshot(@NotNull InWorldRecipeContext context) {
-        ItemCache.ICacheInput input = this.getItem(context);
+        ICacheInput input = this.getItem(context);
         input.shrink(this.item.count());
     }
 
     @Override
     public void rollback(@NotNull InWorldRecipeContext context) {
-        ItemCache.ICacheInput input = this.getItem(context);
+        ICacheInput input = this.getItem(context);
         input.rollbackShrink();
     }
