+++ src/main/java/dev/dubhe/anvilcraft/recipe/anvil/wrap/TimeWarpRecipe.java
@@ -389,6 +389,17 @@ public class TimeWarpRecipe extends AbstractProcessRecipe<TimeWarpRecipe> {
             return this;
         }
 
+        /**
+         * 设置产生量
+         *
+         * @param produce 产量
+         * @return 构建器实例
+         */
+        public Builder produce(int produce) {
+            this.consume(-produce);
+            return this;
+        }
+
         @Override
         protected TimeWarpRecipe of(List<ItemIngredientPredicate> itemIngredients, List<ChanceItemStack> results) {
             return new TimeWarpRecipe(itemIngredients, results, this.hasCauldron.build(), this.produceHeat.build());
